🔧 (zsh): update zprofile and env.Darwin.zsh for improved environment setup
Change the sourced file in .zprofile from os.`uname`.zsh to env.`uname`.zsh
to better reflect the purpose of the file being sourced, which is to set up
the environment.

In env.Darwin.zsh, set the HOMEBREW path based on the architecture, ensuring
compatibility with both ARM and x86_64 systems. This change removes the
dependency on the brew command to determine the prefix, which can fail if
brew is not installed. Instead, it directly sets the path based on the
system architecture.

Additionally, improve the error message when Homebrew is not found by
providing a clear installation command, enhancing user guidance and
experience.